Refactor: Each construct assembles itself
Before, each construct such as a `MadokoHeading` or a `MadokoCodeBlock`
only recognized its first line. The master loop over lines in
`MadokoDocument` was responsible for appending the rest of the lines to
the construct. This led to hard to read code, non-local logic, and in the
end non-extensible code, because the termination conditions are not the
same for all constructs. For example, a paragraph ends with a blank line,
but a fenced code block ends with three back-ticks anywhere on the line.

Refactor so that each construct is responsible for both (1) recognizing
its first line, (2) assembling itself from a sequence of source lines, and
(3) recognizing its terminator.